refactor: ShoppingList 컴포넌트 분리
