+++ src/component/feature/ShoppingListHeader/ShoppingListHeader.styles.ts
@@ -1,0 +1,21 @@
+import { css } from '@emotion/react';
+
+export const headerContainerStyle = css`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  width: 100%;
+  margin-top: 3.6rem;
+  gap: 2.4rem;
+`;
+
+export const filterContainerStyle = css`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+`;
+
+export const filterItemStyle = css`
+  position: relative;
+`;
